OVF: simplify expected return values of transport functions.
Transport functions (transport_iso9660 and transport_vmware_guestinfo)
would return a tuple of 3 values, but only the first was ever used
outside of test.  The other values (device and filename) were just
ignored.

This just simplifies the transport functions to now return content
(in string format) or None indicating that the transport was not found.