Cloud-init: T2309: Added filter to host-name and code cleanup
Since not all data-sources filter hostname in Meta-Data, we need to be sure that the value received from Meta-Data can be applied to the system. The new filter cut all prohibited symbols and takes only that part of the filtered result, which can be used as a hostname.

Additionally, the source was cleaned where it is possible, according to linter recommendations, and more unified with version from 1.3.