Hard code VLAN ID to 0 in set_config_ovf
set_ipaddress was updated to require a VLAN ID parameter in a recent change. set_config_ovf was not updated to reflect the change and it causes cc_vyos.py to fail when booting up a virtual machine created from an OVF template. This change hard codes VLAN ID to 0 to fix the problem for now.