vyos_install: T5220: Copy the whole config folder during installation
In case a user needs to add more material to the configuration (keys, scripts,
etc.) we need to be sure that this material will be copied to a persistent
storage during installation.

This commit replaces copying of only a `config.boot` file to copying the whole
`config` folder, which should satisfy all types of configurations.