cc_vyos: T6007: Run config migrations before Cloud-init
Configuration in `vyos_config_commands` is expected to use a syntax from the
target VyOS version. But, if Cloud-init applies configuration commands to a
default file, all the migrations (from the 0 version to the latest) will still
be executed on top of it.

This process can break proper config, for example, if the current syntax variant
partially repeats one of the older ones, which will trigger a migration without
need.